refactor: sidebar folder structure
